Create a Sphinx project for documentation
The Sphinx project is created in the `docs` subdirectory by entering the
following commands at the top-level directory of this project:

    mkdir docs; cd docs
    sphinx-quickstart -q --ext-autodoc -p Cloudmarker \
                      -a 'Cloudmarker Authors and Contributors'

In this change, no files have been modified after creating the Sphinx
project with the above command. As a result, this change serves as a
reference for the default project created by `sphinx-quickstart`.

Separate changes will be made to customize this project to create
documentation for this project.