Test mock plugin example with unittest.mock.Mock
This way of writing a test works because when we instantiate a
`unittest.mock.Mock` object with keyword arguments during construction,
those keyword arguments appear as attributes of the mock object. Here is
an example:

    >>> import unittest.mock
    >>> m = unittest.mock.Mock(a=3, b=4)
    >>> m.a
    3
    >>> m.b
    4
    >>>

This concept is used in the test changes made in this commit.

In the test, the `util.load_plugin` call does what is done in the Python
snippet above, i.e., instantiate `unittest.mock.Mock` with the
parameters `a=3` and `b=4` as specified in `plugin_config`.

This behaviour of `unittest.mock.Mock` is documented at:
<https://docs.python.org/3/library/unittest.mock.html#unittest.mock.Mock>.
Quoting from the document:

> Mocks can also be called with arbitrary keyword arguments. These will
> be used to set attributes on the mock after it is created. See the
> `configure_mock()` method for details.